Implement multi coin support interfaces
Extends the interfaces to take a coin argument. A coin contains a
cryptocurrency's constants. It also contains the name of a currency to
allow us to check for rule exceptions later on.

Also contains some fixes from review, such as splitting up some switch
cases, splitting up the tests and simplifying the Coin to bitcoinjs-lib
constant mapping function.