Add bitcoin cash and cashaddr
This commit adds code to generate and parse a cashaddress.

The used implementation is a 'typescriptified' version of the old edge
wallet implementation, with some minor patches and improvements. It is
only called from the keymanager, which allows cashaddress use in the
same way as for other addresses.

To faciliate the larger number of tests and test the implementation of
the cashaddress generation directly and through the keymanager, the test
file is split up into multiple smaller files.