Keymanager: Remove cashaddress from address types
To make the calling behavior easier, the cash address types are removed
from the AddressTypeEnum. The code now assumes implicit usage of the
cashaddress format if cashadddr is defined in the coin class. The
cashaddress prefixes are additionally now enforced by choice of the
network (mainnet/testnet) and limited to 'bitcoincash' and 'bchtest'.