Keymanager: Remove path arguments
With the goal of abstracting derivation information away from the
caller, no function should take a 'path' argument. Instead, derivations
are gathered implicitly from BIP43 purpose types and BIP44 coin types.
If a raw seed is passed, it is assumed to be for an airbitz account,
and automatically derives from m/0. Additionally, the BIP44 coin type and
account may be overridden.