Keymanager: Add function for address verification
To allow the caller to learn something about an address before calling
any of the functions, the verifyAddress function is added. It takes the
address and coin name as an argument and returns an enum that is either
'good', for a normal address, 'legacy', for an address that is valid by
legacy rules, or 'bad' for an address that is not valid for that coin.