Keymanager Tests: Move tests to central file
Most keymanager tests are now contained in two files.
`altcointestfixtures` contains the typed test data, while
`keymanagertest` compares the test data with the keymanager's
functionality.

Transaction tests and special case tests, such as creating a replay
protection transaction are not moved, or checking the cash address
derivation, are not moved.

The transaction tests are kept in place. They contain a lot of code
specific to each scenario that might be important for a certain coin, so
little is gained from splitting the test data from the test as well.