UtxoEngine/Processor/Blockbook: Track unconfirmed tx state
To facilitate the tracking of an unconfirmed transaction an additional
CountBase is created to keep track of the unconfirmed transactions. A
count base was chosen to enable extension of its functionality into
tracking non-matured transactions. In such a scenario each coin would
have a maturity limit. This would also make it easier to detect and
track transaction re-orgs.

If a new transaction is saved to the Transaction base with a block
height of 0, its txid is added to count base at index 0.

Upon each incoming block notification all txids in the unconfirmed
transaction count base are fetched. For each of them, their transaction
data is re-fetched from Blockbook and compared with their existing
saved transaction. If its block height is changed, the new transaction
updates the old one and its txid is removed from the count base.

Blockbook does not expose an endpoint for block headers, so it's not
possible to simply check for txid inclusion in them. Instead, the entire
block can be fetched from Blockbook, but it does not seem worth the
extra bandwidth, especially compared to fetching just a couple of
transactions. It might be interesting to eventually fetch the entire
block when reaching a threshold of unconfirmed transactions.