Engine: Add sweep private key functionality
Sweep private key is implemented in the UtxoEngine's sweepPrivateKey
function. To facilitate fetching the to be spent utxos, a temporary
engine state is created. The UtxoEngineState emitter is in turn used as
a callback to notify when all addresses have been scanned. The callback
then resolves the Promise of the sweepPrivateKey function. On success,
it returns an unsigned EdgeTransaction.