Blockbook/Socket: Refactor
Refactors the Socket and Blockbook functions to be more like the
networking code of the old, battle-tested edge-currency-bitcoin plugin.

This commit is meant as pre-cursor work to lay the foundation for a task
queue and multiple sockets.

The new networking code is designed to use fewer callbacks and more
promises and events, favors maps over objects with dynamic keys, has a
more robust keep alive / heartbeat system that takes existing queries
into account, has a cleaner subscription interface and allows for the
potential usage of a task queue and multiple sockets.

Further, some more error handling is introduced, though this will need
further improvement.

Some more test cases in the block book tests are added to test these new
capabilities. Note that this work does not change the interface of the
blockbook functions though.