Network/Engine state: Pick next task
This is an attempt of a port of the pickNextTask routine from the old
edge-currency-bitcoin plugin. An attempt is made to use promises where
possible and execute all networking routines through pickNextTask.

PickNextTask allows the networking and database operations especially
during synching, but also during normal operation, to run completely
async.

PickNextTask is triggered when the socket's pending message queue is
empty enough. It may also be triggered periodically by the timeout in
the main sockets queue.

To keep track of jobs to be done, a cache object is used that is passed
from call to call.

The promises's catch is used to enable full recovery upon a task
failure, meaning allowing the task to be run again in case of a database
or network error.

A wrapper class around the promise, called Deferred, is introduced to
facilitate resolve calls on the promise outside of the Promise
constructor's body. Deferred is additionally templated to give some type
assurances on return.