Cleaners: Implement network cleaner
Adds a cleaner for subscritpion notification messages (new block and new
transaction) and the ServerInfo, Utxos, Transactions and
TransactionBroadcast response messages.

The implementation of the transaction cleaner brought to light that the
current definition of the transaction response type is buggy. It
contained fields not present in the response, and not the complete set
either.

The response for an address message is a mix type, that could not be
easily implemented by the cleaner. Further investigation on its
implementation is required.