Change removeItem util to mutate object passed
Returning a new object from removeItem opens up the potential for race
conditions by changing the reference for a map-like object.
For example, this causes incorrect balances due to race conditions for
UTXO processing using the cache, `taskCashe.processedUtxosCache`.

Changing `removeItem` to `void` instead of an new object, we force
the users to understand that this API is meant to mutate the object
passed and therefore sidestep any object references issues.