Fix setLookAhead infinite loop bug
The `saveAddress` invocation was spreading the `args` parameter from
`setLookAhead` to complete the invocation requirements for
`saveAddress`. This gave way for the opportunity for a bug; `path` was
being overwritten with an unexpected value from the `args` parameter
used by other routines higher in the call-stack.

Replace redundant and confusing `saveAddress` function with first making
a new `IAddress` object (`makeIAddress`) and then simply saving that
object to the processor (using `process.saveAddress`).