Fix txid encoding in signTx
Buffers are turned into Uint8Arrays after crossing the react native
bridge.
The Uint8Array needs to be turned back into a buffer in order for the
toString method to encode it as a hex.
We also opt to reverse the Buffer object instead of the Uint8Array to
avoid the mutation side effect caused by the reverse method on the
Uint8Array.