Fix double balance bug by handling addresses with no utxos
If an address has no utxos, this means that the utxos were spent, and
we must still process the address with an empty set of utxo.
This will cause the utxos to be removed from the processor for the
address and the balance to be decremented by the sum of each utxo's
value.