Deprecate getTotalAddressCount
Use addressSubscribeCache's key length for address count in progress
calculation because it is a sum of distinct addresses.

Remove `totalAddressCount` in in `initializeAddressSubscriptions`
because it's not necessary to check if there are addresses for
initialization.
Just let the function run through all address if they exist because
it's only called once at start up time.