Add generic Blockbook API response types
This is so that we can reuse these generics for other responses.

We also handle and throw the error response in the cleaner rather
than force the API user to handle the error response.
This isolates our server-side error handling to be done where it should,
in a catch statement.