Derive `primaryFormat` with a dedicated `getPrimaryFormat` utility
This implements a single algorithm responsible for getting a wallet's
primaryFormat.
This function uses the existing algorithm for publicKeys and is
now also used in `createPrivateKey`.
This change should be an idempotent, but guarantees our future changes
(about how we select the primary key) is couple to the appropriate
routines.
The goal is to future-proof our existing implementation for when we
enable new wallet formats.

We do not introduce any side-effects to existing wallets (wallets
already persisted prior to any change).