Change scriptPubkey indexed HashBases databases' prefix to 4
Because scriptPubkeys significant bits are after the OP codes, we must
account for those OP code bytes in addition to the 2 hexadecimals of the
scriptPubkey hash.
This means we will have 2 or 4 as the real prefix depending on the
script type for the scriptPubkey.
Prefix of 2 for p2pkh and 4 for p2sh and p2ms.
We don't support p2pk.