Build feeRateUsed for frontend consumption
Computes the sat/vByte used for a given transaction, and adds
`feeRateUsed` object to the `EdgeTransaction` output.

The frontend should be able to render the transaction fee information
for received transactions. However, it only works for new transactions.
To make existing transactions also display the UTXO fees, changes need
to be made in the `edge-core-js` repository to modify the cache fetching
and merging logics.