Mark or remove UTXOs as spent on saveTx/network event
This consolidates the logic of processing the UTXOs in a single
processor function in the engine state.
This is better for maintainability.

We mark UTXOs as spent until they're confirmed, after which we remove
them from the local database.

We shouldn't allow UTXOs to be market as unspent after it has been
marked as spent due to network state inconsistency which may occur.
Instead, we should do this in a future dropped transaction algorithm.