Remove replay-protection hack that's didn't make sense
We cannot change the format of the addresses to bip49 if the wallet
does not support this format because it doesn't have an xpriv in its
walletInfo.

Instead, we will use the deriviationLevelScriptHash function to detect
the correct P2SH script type for the address even if the format is
bip44.