Fix race condition causing inconsistency with new UTXOs in UTXO-set
Blockbook servers wont report new UTXOs until some time after broadcast
of the transaction. After a new transaction is broadcast and saved,
the UTXO is added and then quickly removed after querying for the
address' UTXOs during address the subscription routine.

To solve this issue, we remove the query for address UTXOs that have
an existing Blockbook address subscription. This introduces a
bug case where UTXOs may be missed during the narrow window of
of disconnection when doing address re-subscription.