Replace NumbWalletInfo primaryFormat with privateKeyFormat
Now that we've made a strict definition for the private key format, we
should use this as it is more closely coupled with the key-formats
spec. There no longer is an internal "primary format" concept that was
determined by the ordering of formats in then currency's EngineInfo.
Furthermore, there is no notion of a "default format from which to
derive addresses".