Add serverConfig to EngineInfo with fallback servers
This infra is similar to accountbased currencies in that we maintain
a list of server configs that can match with a specific key from
initOptions.
The only difference with this implementation is that there is no
info-server support for this integration yet.
These servers will only be used as fallback servers if WebSocket
broadcasts fail.