Rename files to match main exported interface
The convention is such that if a file's main purpose is to define a
interface/type by a class or factory function, then the file name
should match the name of that interface.

So, we remove "make" from file names, and we make some changes to
interface type names to match the file name.