Make sequence field required for keymanager UTXO type
This UTXO type is the data type we use when communicating with the
keymanager library which is a layer of abstraction that sits on top
of the AltcoinJS library. It is used in making and signing transaction
and therefore should always have access to the sequence field for all
inputs.

Although poorly named, the UTXO type is both the output and input types
for the keymanager. Perhaps in the future this could be better organized
but for now, UTXO is the type we use for tx inputs and outputs when
communicating with the AltcoinJS library.