Replace `userSettings` with `serverCache` and parameter passing
By relying on `serverCache` exclusively for our most up-to-date state
when calling `refreshServers` we can ensure that our initial load time
and subsequent calls are using the correct state for the wallet.

This fixes a race condition during initial load time: `userSettings`
state was the set to `defaultSettings`. Now that we persists the last
user's settings in the `serverCache` under the `'customServers'` key,
we can use that instead.

We can eliminate the need for a `userSettings` instance variable by
using a parameter to pass the new customer servers when invoking
`refreshServers`. It never made sense to rely on instance variables, or
properties, to pass state around. This new approach is the correct way
to manage state.