Extract needsTxSpecific to a function
There is no point to calculating this on every pickNextTask.
This change simplifies the processor arguments and removes unnecessary
CPU usage on every pickNextTask. More importantly, it removes the extra
memory usage by allocating a boolean to be GC'd after pickNextTask pops
from the stack.
All in all, this change seems like a win.