Fix checkpoint initialization on upgrade
We were only updating the checkpoint in response to incoming transactions, which doesn't happen if the phone is already synced. The next incoming transaction after the upgrade will encounter and `undefined` checkpoint, and won't generate a notifiation. Instead, also update the checkpoint when the sync finishes, to ensure that existing accounts will detect incoming funds without needing a full resync.