Replace secp256k1 library with bigint compatible version
In addition to replacing the library, we're changing how the ECPair library is first instantiated. It's wrapped in a function so that any errors thrown during `initEccLib` can bubble up nicely instead of crashing on launch.

Dependencies of edge-currency-plugins, like altcoinjs, already depend on on bigint support so this is not considered a breaking change.